Archive + Mappool page are partially working again
- I haven't test with the database to see if the CRUD function still
  working normally or I have to adjust them. This will be done on next
  commit.

- Adjust naming for JSON files match with this project's naming
  convention.

- Also update file naming convention (Part 3)
